requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==1.4.53
passlib==1.7.4
bcrypt==4.0.1
python-jose==3.3.0
python-multipart==0.0.6
pillow>=10.2.0
httpx==0.25.2
openai==1.3.8
google-cloud-aiplatform==1.38.0
google-auth==2.23.4
google-api-python-client==2.108.0
python-dotenv==1.0.0
email-validator==2.1.0
pyzbar==0.1.9
transformers>=4.21.0
torch>=1.12.0
torchvision>=0.13.0
requests>=2.28.0